feat(payment): INT-1479 Modify payment strategy
