--- src/payment/strategies/cybersource/cybersource-payment-processor.ts
@@ -1,39 +1,0 @@
-import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
-import {
-    OrderActionCreator,
-    OrderPaymentRequestBody,
-    OrderRequestBody
-} from '../../../order';
-import { OrderFinalizationNotRequiredError } from '../../../order/errors';
-import PaymentActionCreator from '../../payment-action-creator';
-import { PaymentRequestOptions } from '../../payment-request-options';
-
-import { CreditCardInstrument } from '../../payment';
-import PaymentMethod from '../../payment-method';
-
-export default class CyberSourcePaymentProcessor {
-    constructor(
-        private _store: CheckoutStore,
-        private _orderActionCreator: OrderActionCreator,
-        private _paymentActionCreator: PaymentActionCreator
-    ) {}
-
-    initialize(paymentMethod: PaymentMethod): Promise<InternalCheckoutSelectors> {
-        return Promise.resolve(this._store.getState());
-    }
-
-    execute(payment: OrderPaymentRequestBody, order: OrderRequestBody, paymentData: CreditCardInstrument, options?: PaymentRequestOptions): Promise<InternalCheckoutSelectors> {
-        return this._store.dispatch(this._orderActionCreator.submitOrder(order, options))
-            .then(() =>
-                this._store.dispatch(this._paymentActionCreator.submitPayment({ ...payment, paymentData }))
-            );
-    }
-
-    finalize(options?: PaymentRequestOptions): Promise<InternalCheckoutSelectors> {
-        return Promise.reject(new OrderFinalizationNotRequiredError());
-    }
-
-    deinitialize(options?: PaymentRequestOptions): Promise<InternalCheckoutSelectors> {
-        return Promise.resolve(this._store.getState());
-    }
-}
