+++ src/payment/strategies/cybersource/index.ts
@@ -2,5 +2,3 @@ export * from './cardinal';
 
 export { default as CyberSourcePaymentStrategy } from './cybersource-payment-strategy';
 export { default as CardinalScriptLoader } from './cardinal-script-loader';
-export { default as CyberSourceThreeDSecurePaymentProcessor } from './cybersource-threedsecure-payment-processor';
-export { default as CyberSourcePaymentProcessor } from './cybersource-payment-processor';
